requirements.txt
dlib
opencv-python
torch
torchvision
flask
boto3
botocore
cycler
iopath
jmespath
kiwisolver
matplotlib
pillow
s3path
s3transfer
smart-open
smart-path
torchgpipe
uuid
torchvision
requests
fvcore
fire
psycopg2
python-dotenv
flask-cors